### 2025-02-19 #7648 @alfredohernandezinostroza
@Kevin-Mattheus-Moerman Hello! I just wanted to follow up on this resubmission, we updated the paper according to what we talked here: [this issue](https://github.com/openjournals/joss-reviews/issues/6626#issuecomment-2236180181). Thanks!
