### 2025-02-20 #7648 @Kevin-Mattheus-Moerman
# Notes to editors:

* This is a re-submission of:
https://github.com/openjournals/joss-reviews/issues/6626

* The branch used for the paper is not for the main project. I ran `cloc` locally for the main branch which produced:
```
Language                     files          blank        comment           code
SVG                              1              0              0           2671
HTML                            17            367             51           2149
Python                          20            331            326           1968
JavaScript                      12            138            228            911
CSS                              4            190             35            780
Text                             3             32              0            185
Markdown                         3             82             32            154
TeX                              1             16              0            141
reStructuredText                15            372            277            101
CSV                              2              0              0             64
JSON                             2              2              0             54
TOML                             1              3              0             45
DOS Batch                        1              8              1             26
YAML                             1              3              0             10
make                             1              4              7              9
SUM:                            84           1548            957           9268
```
