### 2025-03-06 #7648 @alfredohernandezinostroza
@Kevin-Mattheus-Moerman Hello! I was considering suggesting reviewers for the submission. Is it mandatory for them to be on the [list of reviewers](https://reviewers.joss.theoj.org/reviewers)? Thanks!
