### 2025-04-05 #7648 @majensen
@chryswoods would you be willing to review [this work](#7648) for [JOSS](https://joss.theoj.org)? I appreciate your consideration! --@majensen
