### 2025-04-05 #7648 @majensen
Hi @chrisamiller, would you have time and inclination to review [this work](#7648) for [JOSS](https://joss.theoj.org)? I'd welcome your expertise on this subject. Thanks! --@majensen
