### 2025-04-17 #7648 @majensen
@davidnsousa - would you be willing to review [this work](#7648) for [JOSS](https://joss.theoj.org)? I would really appreciate your expertise. Thanks for considering —@majensen
