### 2025-04-25 #8092 @tushardave26
Hello, @alfredohernandezinostroza

Thank you for submitting your work to JOSS and sharing the MetaSpread software with the community. I appreciate the effort and thought you put into developing this tool for modeling cancer growth and metastasis.

As I begin my review, I have a few initial comments and suggestions:

**Authorship Clarification:**
Could you please describe the contributions of the second author listed on the paper? I do not see these authors' commits or contributions in the GitHub repository.

**README.md improvements:**
- Please provide a complete list of dependencies or requirements for MetaSpread, including the Python version you have tested and any additional Python packages required to run the software.
- While the documentation includes some usage examples, I recommend adding a few basic usage examples directly to the README.md. This would help showcase your software's capabilities and make it easier for new users to get started.

**Interactive mode Warning**
When running the `MetaSpread` in interactive mode, I received the following warning message:
```bash
sh: cls: command not found
```
This occurs because `os.system('cls')` is used in `interactive.py` to clear the terminal/screen, which only works on Windows. The Unix/Linux systems use the `clear` command for the same purpose. For better cross-platform compatibility, I recommend replacing this with:
```python
import os

def clear_screen():
    if os.name == 'nt':
        os.system('cls')
    else:
        os.system('clear')
```

**Recommendation for Documentation Imporvement**
1. Provide a Complete Workflow Example:
    - Add a step-by-step, real or simulated example that starts from setting up simulation parameters, running a simulation, post-processing, generating figures, and interpreting outputs.
    - Show actual command lines and sample outputs (e.g., directory structure, CSV snapshots, sample plots).
    - Ideally, walk through a sample scenario from start to finish, including input files and the resulting figures and outputs.
2. Show Output at Each Step:
    - For each workflow step (simulation, data analysis, graphics, videos), show what files are produced and include example visualizations or screenshots.
    - Help users understand what to expect and verify successful completion.
3. Improve Navigation for New Users:
    - Consider adding a "Quickstart" section with the minimal steps needed to run a first simulation and generate a figure.
    - Link out to more advanced or detailed sections for users who want to dive deeper.

**State of the Field – Comparative Context in the Paper**
While the paper provides a solid background and includes references to related work, I noticed that there is no direct comparison between MetaSpread and other commonly used cancer simulation packages, such as CancerSim or similar agent-based modeling frameworks. Including a short section or paragraph summarizing the unique features or advantages of MetaSpread, as well as how it complements or differs from existing tools, would help readers better understand its value and place within the broader field. This context would be beneficial for new users evaluating which tool to adopt for their research.

I’d like to request that you incorporate all the suggested code changes and release a new version of the package. I will then be able to test its functionality again.

For context, I have attempted to get the interactive mode working on my MacBook Pro but have not been successful. Additionally, while the CLI version works for some steps, it does not work for all functionalities. Providing a new release with the recent fixes and improved documentation will help me (and other users) to properly test the complete workflow and give more specific feedback.

Thank you!
