### 2025-04-25 #8092 @chryswoods
I agree with @tushardave26 - please create a new release with the requested fixes and improved documentation and I will review that. I also use a MacBook Pro, and would need to be able to properly test the complete workflow to be able to complete a review.

I have managed to run the code, but also experience issues like @tushardave26. These are, I think, because this was developed on Windows, and has some non-portable code that won't work on Mac or Linux.

In addition to the use of `cls`, the code looks for files using a `\` backslash directory separator. On Mac and Linux a `/` forward slash is used.

This leads to errors in analysis, e.g. running the simulation works

```
$ python -m metaspread
Configs file`for new simulation "simulations_configs.csv" not found! Creating a configs file with default values.
sh: cls: command not found
==================================================================
Welcome to MetaSpread: a Cancer Simulation Program!
==================================================================
➤ New simulation
   Load simulation
   Postprocessing
   Exit

sh: cls: command not found
==================================================================
New simulation
==================================================================
sh: cls: command not found
Select the total timesteps the simulation will take: sh: cls: command not found
Select the total timesteps the simulation will take: 100
sh: cls: command not found
Select the size of the intervals for which the information will be collected: 1

         Folder for this simulation: Simulations/Sim-max_steps-100-collection_period-1-cells-388-grids_number-3
         Saving agents data at: Simulations/Sim-max_steps-100-collection_period-1-cells-388-grids_number-3
         Saving Mmp2 data at: Simulations/Sim-max_steps-100-collection_period-1-cells-388-grids_number-3/Mmp2
         Saving Ecm data at: Simulations/Sim-max_steps-100-collection_period-1-cells-388-grids_number-3/Ecm
         Saving Vasculature data at: Simulations/Sim-max_steps-100-collection_period-1-cells-388-grids_number-3/Vasculature
         Saving all the simulations parameters at: Simulations/Sim-max_steps-100-collection_period-1-cells-388-grids_number-3/configs.csv
Starting simulation from zero!
Finished the simulation at time step 100!
```

but then analysis fails because it can't find the output files

```
$ python -m metaspread
sh: cls: command not found
==================================================================
Welcome to MetaSpread: a Cancer Simulation Program!
==================================================================
➤ New simulation
   Load simulation
   Postprocessing
   Exit
^[[Bsh: cls: command not found
==================================================================
Welcome to MetaSpread: a Cancer Simulation Program!
==================================================================
   New simulation
➤ Load simulation
   Postprocessing
   Exit
^[[Bsh: cls: command not found
==================================================================
Welcome to MetaSpread: a Cancer Simulation Program!
==================================================================
   New simulation
   Load simulation
➤ Postprocessing
   Exit

sh: cls: command not found
['Sim-max_steps-100-collection_period-1-cells-388-grids_number-3']
Traceback (most recent call last):
  File "<frozen runpy>", line 198, in _run_module_as_main
  File "<frozen runpy>", line 88, in _run_code
  File "~/.pyenv/versions/3.11.9/lib/python3.11/site-packages/metaspread/__main__.py", line 23, in <module>
    main_menu()
  File "~/.pyenv/versions/3.11.9/lib/python3.11/site-packages/metaspread/interactive.py", line 24, in main_menu
    postprocessing_menu()
  File "~/.pyenv/versions/3.11.9/lib/python3.11/site-packages/metaspread/interactive.py", line 59, in postprocessing_menu
    folder_names = ["Exit"] + get_folder_names(directory_path)
                              ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "~/.pyenv/versions/3.11.9/lib/python3.11/site-packages/metaspread/interactive.py", line 161, in get_folder_names
    folder_names.sort(key=lambda x: os.path.getmtime(f"{directory_path}\{x}"), reverse=True)
  File "~/.pyenv/versions/3.11.9/lib/python3.11/site-packages/metaspread/interactive.py", line 161, in <lambda>
    folder_names.sort(key=lambda x: os.path.getmtime(f"{directory_path}\{x}"), reverse=True)
                                    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "<frozen genericpath>", line 55, in getmtime
FileNotFoundError: [Errno 2] No such file or directory: 'Simulations\\Sim-max_steps-100-collection_period-1-cells-388-grids_number-3'
```

This directory does exist

```
$ ls Simulations/Sim-max_steps-100-collection_period-1-cells-388-grids_number-3
CellsData.csv                  configs.csv                    Ecm                            Mmp2                           Time when grids were populated Vasculature
```

I recommend using [`os.path.join`](https://docs.python.org/3/library/os.path.html#os.path.join) when creating paths to files and directories, as this will make sure to use the right directory separator for the operating system.
