### 2025-04-25 #8092 @tushardave26
@chryswoods – Thank you for sharing the traceback. I’ve noticed that the path issue you mentioned occurs only when running `postprocess` in interactive mode. When executed as a CLI command, `postprocess` runs without any problems. Please see the log below from my CLI execution for reference.

I agree that using os.path.join is the right approach to resolve cross-platform compatibility issues. In fact, I’ve noticed that @alfredohernandezinostroza is already using it in several places in the codebase, but it appears that it hasn’t been applied consistently throughout all instances.
```bash
$ python -m metaspread postprocess data Sim-max_steps-10-collection_period-25-cells-388-grids_number-3
	Analyzing data in the folder Simulations/Sim-max_steps-10-collection_period-25-cells-388-grids_number-3

Using cells data at: Simulations/Sim-max_steps-10-collection_period-25-cells-388-grids_number-3/CellsData.csv
Using Ecm data in the folder: Simulations/Sim-max_steps-10-collection_period-25-cells-388-grids_number-3/Ecm
Using Mmp2 data in the folder: Simulations/Sim-max_steps-10-collection_period-25-cells-388-grids_number-3/Mmp2
Using vasculature data at: Simulations/Sim-max_steps-10-collection_period-25-cells-388-grids_number-3/Vasculature
Warning: the run for this simulation terminated early
Max step reached is 10 while 10 was expected.

Saving tumor data in the folder: Simulations/Sim-max_steps-10-collection_period-25-cells-388-grids_number-3/Data analysis/Tumor dynamics
Saving cells numbers data in the folder: Simulations/Sim-max_steps-10-collection_period-25-cells-388-grids_number-3/Data analysis/Cells growth
Saving vasculature data in the folder: Simulations/Sim-max_steps-10-collection_period-25-cells-388-grids_number-3/Data analysis/Vasculature dynamics

Grid: 1
	Saving tumor data...
	Saving cells numbers graph data...

Grid: 2
	Saving tumor data...
	Saving cells numbers graph data...

Grid: 3
	Saving tumor data...
	Saving cells numbers graph data...
Saving vasculature...
```
Since the previous step ran fine for me, I proceeded to the next step in the workflow to generate the graphics. However, I encountered a different error when running the `postprocess graphics` command via the CLI. Below is the error traceback I received.
```bash
python -m metaspread postprocess graphics Sim-max_steps-10-collection_period-25-cells-388-grids_number-3
~/Peer Review/MetaSpread/venv/lib/python3.12/site-packages/metaspread/__main__.py:43: UserWarning: No amount of pictures given to graphics. Generating all pictures.
  warnings.warn("No amount of pictures given to graphics. Generating all pictures.")
Analyzing data in the folder Simulations/Sim-max_steps-10-collection_period-25-cells-388-grids_number-3

Using cells data at: Simulations/Sim-max_steps-10-collection_period-25-cells-388-grids_number-3/CellsData.csv
Using Ecm data in the folder: Simulations/Sim-max_steps-10-collection_period-25-cells-388-grids_number-3/Ecm
Using Mmp2 data in the folder: Simulations/Sim-max_steps-10-collection_period-25-cells-388-grids_number-3/Mmp2
Using vasculature data at: Simulations/Sim-max_steps-10-collection_period-25-cells-388-grids_number-3/Vasculature
Warning: the run for this simulation terminated early
Max step reached is 10 while 10 was expected.

Saving tumor images in the folder: Simulations/Sim-max_steps-10-collection_period-25-cells-388-grids_number-3/Graphical analysis/Tumor dynamics
Saving Mmp2 images in the folder: Simulations/Sim-max_steps-10-collection_period-25-cells-388-grids_number-3/Graphical analysis/Mmp2 dynamics
Saving Ecm images in the folder: Simulations/Sim-max_steps-10-collection_period-25-cells-388-grids_number-3/Graphical analysis/Ecm dynamics
Saving cells numbers images in the folder: Simulations/Sim-max_steps-10-collection_period-25-cells-388-grids_number-3/Graphical analysis/Cells growth
Saving vasculature images in the folder: Simulations/Sim-max_steps-10-collection_period-25-cells-388-grids_number-3/Graphical analysis/Vasculature dynamics
Saving histogram images in the folder: Simulations/Sim-max_steps-10-collection_period-25-cells-388-grids_number-3/Graphical analysis/Positions histogram
Saving histogram image in the folder: Simulations/Sim-max_steps-10-collection_period-25-cells-388-grids_number-3/Graphical analysis/Radius and diameter

Grid: 1
	Plotting Mmp2 graphs...
	Plotting Ecm graphs...
	Plotting tumor graphs...
	Plotting histogram graphs...
	Plotting cells numbers graph...

Grid: 2
	Plotting Mmp2 graphs...
	Plotting Ecm graphs...
	Plotting tumor graphs...
	Plotting histogram graphs...
	Plotting cells numbers graph...

Grid: 3
	Plotting Mmp2 graphs...
	Plotting Ecm graphs...
	Plotting tumor graphs...
	Plotting histogram graphs...
	Plotting cells numbers graph...
Plotting vasculature...
Plotting radius and diameter history graph...
Traceback (most recent call last):
  File "<frozen runpy>", line 198, in _run_module_as_main
  File "<frozen runpy>", line 88, in _run_code
  File "~/Peer Review/MetaSpread/venv/lib/python3.12/site-packages/metaspread/__main__.py", line 44, in <module>
    graphgenerator.generate_graphs(simulation_folder, amount_of_pictures)
  File "~/Peer Review/MetaSpread/venv/lib/python3.12/site-packages/metaspread/graphgenerator.py", line 385, in generate_graphs
    plot_radius_diameter_history(radius_history_df, radius_diameter_images_path, step, real_delta_time)
                                                                                 ^^^^
UnboundLocalError: cannot access local variable 'step' where it is not associated with a value
```

@alfredohernandezinostroza - I’d like to reiterate my previous request: please incorporate all the suggested changes, generate a new package, and share it for testing. I also recommend testing the updated package on a Linux system to ensure broader compatibility.
