### 2025-05-07 9282499
add requiremebnts to readme
