### 2025-05-07 a5ada6a
fix clear screen not working in OS that are not Windows
