### 2025-05-07 e7bdc4e
make get_folder_names OS agnostic
