+++ paper.bib
@@ -32,6 +32,16 @@
   doi={10.1016/j.jocs.2019.101067 }
 }
 
+@article{cancersim, doi = {10.21105/joss.02436},
+  url = {https://doi.org/10.21105/joss.02436},
+  year = {2020},
+  publisher = {The Open Journal},
+  volume = {5},
+  number = {53},
+  pages = {2436},
+  author = {Luka Opasic and Jacob G. Scott and Arne Traulsen and Carsten Fortmann-Grote},
+  title = {CancerSim: A Cancer Simulation Package for Python 3},
+  journal = {Journal of Open Source Software} } 
 
 @misc{documentation,
 	title = {MetaSpread 2024.0.1 documentation},
