+++ paper.bib
@@ -74,17 +74,6 @@
   doi={10.1007/s11517-022-02514-2}
 }
 
-@article{opasic2020cancersim,
-  title={CancerSim: A Cancer Simulation Package for Python 3},
-  author={Opasic, Luka and Scott, Jacob G and Traulsen, Arne and Fortmann-Grote, Carsten},
-  journal={Journal of Open Source Software},
-  volume={5},
-  number={53},
-  pages={2436},
-  year={2020},
-  doi={10.21105/joss.02436}
-}
-
 @inproceedings{python-mesa-2020,
   title={MESA: an agent-based modeling framework},
   author={Masad, David and Kazil, Jacqueline},
