+++ paper.md
@@ -36,7 +36,7 @@ We develop and provide MetaSpread, an open-source simulation package and interac
 
 Models of tumor growth and metastatic spread are critical for understanding the key underlying biological processes and clinical evolution in patients. Mathematical models can be of different level of detail, computational or theoretical, spatial or non-spatial in nature, and can have several mechanisms explicit or implicitly embedded in them, including interaction with resources, biomechanical signals, cellular competition, mutation and migration [@waclaw_spatial_2015;@franssen2021novel;@macnamara2020computational;@chaplain2020multiscale;@cancersim]. While theoretical and analytical advances remain crucial in mathematical models of cancer, computational approaches that offer direct simulation platforms for efficient numerical exploration, focused study and hypothesis testing are also very much needed. Here, we contribute to this aspect, by offering an open source simulation framework in Python for spatio-temporal progression of tumor and metastatic spread. We build the simulation framework on a hybrid mathematical model developed by @franssen2019, extending the previous work [@anderson1998continuous], in close agreement with empirical data [@sabeh2009protease; @newton2015spatiotemporal]. This contribution aims to bridge the gap between mathematicians, oncologists, biologists, computer scientists and interested researchers working in the field of cancer metastatic progression.
 
-# Comparison with Existing Cancer Simulation Tools
+# Comparison with existing cancer simulation software
 
  There are other packages that simulate cancer processes in the literature. For example, CancerSim [@cancersim] is a software that simulates somatic evolution of tumors. It produces virtual spatial tumours with variable extent of intratumour genetic heterogeneity and realistic mutational profiles. By applying multi-region sampling to simulated tumors to obtain mutation profiles, one can later match them with sequencing data. This makes CancerSim a useful tool for studying various sampling strategies in clinical cancer diagnostics. Although there are some parallels of CancerSim with our software MetaSpread, like the agent-based formulation for individual cells, there are also several differences. For example, CancerSim does not address metastatic processes, nor does it account explicitly for the cancer cell interaction with the extracellular matrix or MMP2, or the specificities of cell movement like diffusion and haptotaxis that our hybrid model based on [@franssen2019] includes.  The objectives being different, CancerSim focused on cell heterogeneity and evolution processes with their effect on tumor growth, and MetaSpread focused on cell population processes (deterministic and stochastic) leading to metastatic spread. We believe both of these frameworks can be complementary, and used in parallel or combined to study and integrate distinct aspects of cancer dynamics.  
 
