### 2025-05-14 3986ddb
fix graphical bug on main menu when at the begginning or the end of the list
