+++ paper.md
@@ -36,10 +36,6 @@ We develop and provide MetaSpread, an open-source simulation package and interac
 
 Models of tumor growth and metastatic spread are critical for understanding the key underlying biological processes and clinical evolution in patients. Mathematical models can be of different level of detail, computational or theoretical, spatial or non-spatial in nature, and can have several mechanisms explicit or implicitly embedded in them, including interaction with resources, biomechanical signals, cellular competition, mutation and migration [@waclaw_spatial_2015;@franssen2021novel;@macnamara2020computational;@chaplain2020multiscale;@cancersim]. While theoretical and analytical advances remain crucial in mathematical models of cancer, computational approaches that offer direct simulation platforms for efficient numerical exploration, focused study and hypothesis testing are also very much needed. Here, we contribute to this aspect, by offering an open source simulation framework in Python for spatio-temporal progression of tumor and metastatic spread. We build the simulation framework on a hybrid mathematical model developed by @franssen2019, extending the previous work [@anderson1998continuous], in close agreement with empirical data [@sabeh2009protease; @newton2015spatiotemporal]. This contribution aims to bridge the gap between mathematicians, oncologists, biologists, computer scientists and interested researchers working in the field of cancer metastatic progression.
 
-# Comparison with existing cancer simulation software
-
- There are other packages that simulate cancer processes in the literature. For example, CancerSim [@cancersim] is a software that simulates somatic evolution of tumors. It produces virtual spatial tumours with variable extent of intratumour genetic heterogeneity and realistic mutational profiles. By applying multi-region sampling to simulated tumors to obtain mutation profiles, one can later match them with sequencing data. This makes CancerSim a useful tool for studying various sampling strategies in clinical cancer diagnostics. Although there are some parallels of CancerSim with our software MetaSpread, like the agent-based formulation for individual cells, there are also several differences. For example, CancerSim does not address metastatic processes, nor does it account explicitly for the cancer cell interaction with the extracellular matrix or MMP2, or the specificities of cell movement like diffusion and haptotaxis that our hybrid model based on [@franssen2019] includes.  The objectives being different, CancerSim focused on cell heterogeneity and evolution processes with their effect on tumor growth, and MetaSpread focused on cell population processes (deterministic and stochastic) leading to metastatic spread. We believe both of these frameworks can be complementary, and used in parallel or combined to study and integrate distinct aspects of cancer dynamics.  
-
 # Cancer growth and spread model
 
 A 2-dimensional multigrid hybrid spatial model of cancer dynamics is developed in Python (see Figure 1 for a snapshot illustration), where the first grid represents the primary tumor site, and the secondary grids represent metastatic sites, connected via the vasculature compartment. Here we combine the stochastic individual based dynamics of single cells with deterministic dynamics of the abiotic factors. In the tumor site we consider two different cancer cell phenotypes: epithelial (epithelial-like) and mesenchymal (mesenchymal-like) cells. The epithelial-like (E) cancer cells reproduce at a higher rate, but diffuse more slowly than mesenchymal (M) cells, which reproduce at a lower rate but diffuse more rapidly. Furthermore, epithelial cells require the presence of mesenchymal cells to be able to intravasate into normal vessel entry-points. The exception to this are ruptured vessels, that allow for the intravasation of any type of cancer cell. The cellular growth and movement in space is modeled considering 2 partial differential equations, where diffusion and haptotaxis movement are implemented. The model includes two additional equations: one for the spatio-temporal dynamics of matrix metalloproteinase 2 (MMP-2), a chemical that favors the spread of cancer cells, and another for the degradation of the extracellular matrix (ECM), which also favors the haptotactic movement of the cancer cells. 
@@ -198,9 +194,13 @@ The parameters non-dimensional values, as well as their code equivalent name are
 
[…]
