### 2025-05-15 #8092 @alfredohernandezinostroza
Hello @tushardave26 @chryswoods! We have already improved several aspects according to your feedback:

Done:
- Paper update with a comparison with CancerSim.
- Added python and package requirements to README.md.
- New version available, fixing the bugs you mentioned. The testing was done both in Windows and Unix systems. You can install it with ```pip```.

To do:

- Add basic usage to README.md
- Documentation Improvement

Additionally, here are the details on the authors contributions:

- Authorship Clarification:

> **Alfredo Hernandez-Inostroza:**
> Conceptualization,
> Investigation,
> Software development,
> Validation,
> Visualisation,
> Writing - Original Draft Preparation,
> Writing - Review and Editing
>
> **Erida Gjini:**
> Conceptualization,
> Investigation,
> Methodology,
> Project administration,
> Supervision,
> Writing - Original Draft Preparation,
> Writing - Review and Editing

Thank you for your comments!
