### 2025-06-09 alfredohernandezinostroza/MetaSpread#1 @tushardave26
Missing CancerSim comparison in the paper
The issue is created as part of the JOSS review process for openjournals/joss-reviews#8092.

> * Paper update with a comparison with CancerSim.

I have regenerated the Paper, as you can see [here](https://github.com/openjournals/joss-reviews/issues/8092#issuecomment-2956974522), but I can't see the updates that you made. Perhaps @majensen can assist you in adding the updates.
