+++ docs/source/introduction.rst
@@ -1,0 +1,337 @@
+Introduction
+======================================
+
+.. contents::
+   :depth: 2
+
+Summary
+-------
+
+MetaSpread is an open source simulation package and interactive program in Python for tumor growth and metastatic spread, based on a mathematical model by :cite:t:`franssen2019`. This paper proposed a hybrid modeling and computational framework where cellular growth and metastatic spread are described and simulated in a spatially explicit manner, accounting for stochastic individual cell dynamics and deterministic dynamics of abiotic factors. This model incorporates several key processes such as the growth and movement of epithelial and mesenchymal cells, the role of the extracellular matrix, diffusion, haptotaxis, circulation and survival of cancer cells in the vasculature, and seeding and growth in secondary sites. In the software that we develop, these growth and metastatic dynamics are programmed using MESA, a Python Package for Agent-based modeling :cite:p:`python-mesa-2020`.
+
+Installation
+------------
+
+MetaSpread is available as a official PyPI package. To install, you will need to have PIP installed. Afterwards, run:
+
+.. code-block:: console
+   
+   pip install metaspread
+
+or:
+
+.. code-block:: console
+
+   python -m pip install metaspread
+
+Manual installation can be done by downloading the repo.
+
+Finally, the program can be run interactively with:
+
+
+.. code-block:: console
+
+   python -m metaspread
+
+Or it can be run purely through command line arguments, as detailed in the following section.
+
+Usage
+-----------------
+
+The program can be run both interactively through the command line, or with explicit user command line arguments.
+
+
+.. image:: main_menu.png
+When run interactively, starting from the main menu, the following possibilities are offered: 
+
+- **Run a new simulation:** the user can choose the *New Simulation* option to run a new simulation, with the arguments to be specified by the user being the maximal time for the dynamics, and the frequency of saving data (temporal resolution). Any other simulation parameter (see  Table :ref:`Table <table-sim-parameters>` ) will be taken from the *simulation\_configs.csv* file in the main folder. At the end of the simulation the dynamics of the grids, including agents (cells and vasculature points), the vasculature dynamics and the MMP2 and ECM are saved in a properly identified directory, including a *configs.csv* recording the used parameters for this particular simulation. The file *CellsData.csv* in this directory will include all the information of all cells and vasculature points in the simulation, for every time step.
+
+  - The *simulation_configs.csv* file can be modified both in a code editor, or in a spreadsheet processing software, such as Microsoft Excel:
+
+  .. figure:: csv_both.png
+  The *simulation_configs.csv* file in Visual Studio Code (left) and in Microsoft Excel (right).
+..   .. figure:: csv_code.png
+..   The *simulation_configs.csv* file in Visual Studio Code.
+  .. figure:: csv_excel.png
+..   The *simulation_configs.csv* file in Microsoft Excel.
+  
+  - In addition, in the ECM and MMP2 folders there will be files containing the values of these factors for each time step, not requiring any postprocessing.
+  
+  - The vasculature folder will contain several *.json* files with the state of the vasculature at each time step. That is, they will contain a dictionary showing the clusters that were present at each time step. Further information can be extracted by using the **data analysis** option.
+  
+  - The folder *Time when grids got populated* will have a file that will simply show the time step for which each grid (primary or secondary site) got populated.
+
[…]

[…]
