### 2025-07-21 f58a875
add more checks for correct simulation_configs input
