+++ docs/source/tutorial-interactive.rst
@@ -23,12 +23,14 @@ or:
 Running metaspread
 ------------------
 
-Once installed, create a folder for this tutorial project. For example, this command will create a folder called "tutorial-interactive":
+Once installed, create a directory for this tutorial called "tutorial-interactive":
+
 .. code-block:: console
 
    mkdir tutorial-interactive
 
-Then, navigate to this folder:
+Then, navigate to the directory:
+
 .. code-block:: console
 
    cd tutorial-interactive
@@ -53,35 +55,70 @@ First, we will focus on running a new simulation.
 Running a Simulation
 --------------------
 
-1. **Editing the parameters**: the default parameters are already set in the ``simulation_configs.csv`` file. You can edit this file to change the parameters of your simulation. To edit the file, you can use any text editor or spreadsheet software that supports CSV format. For this tutorial, we will change the default cell number to 400, the number of grids to 2. For simplicity, we will use a text editor. The file should look like this:
+1. **Editing the parameters**: the default parameters are already set in the ``simulation_configs.csv`` file. You can edit this file to change the parameters of your simulation. To edit the file, you can use any text editor or spreadsheet software that supports CSV format. For this tutorial, we will change the number of grids to 2. For simplicity, we will use a text editor. The file should look like this:
 
-.. image:: notepad_parameters.png
+   .. image:: notepad_parameters.png
 
    In principle, in order to change the amount of initial cells, and the number of grids, it is only necessary to change the values in the ``cells`` and ``grids_number`` rows, respectively. 
 
-.. image:: changing_parameters_old.png
+   .. image:: changing_parameters_old.png
 
    However, this will give an error when running the simulation:
 
-.. image:: error_grids_number.png
+   .. image:: error_grids_number.png
 
-   This is because there are certain parameters that depend on the number of grids, such as the ``extravasation_probs`` and ``secondary_sites_vessels``. Thus, it is necessary to change these parameters as well. The ``extravasation_probs`` parameter should be a list of probabilities for each secondary grid, and the ``secondary_sites_vessels`` parameter should be a list of the number of vessels in each secondary grid.Both of these parameters should have a length equal to the number of grids - 1. Thus, we can change the parameters accordingly:
-   
-.. image:: changing_parameters.png
+   This is because there are certain parameters that depend on the number of grids, such as the ``extravasation_probs`` and ``secondary_sites_vessels``. Thus, it is necessary to change these parameters as well. The ``extravasation_probs`` parameter should be a list of probabilities for each secondary grid that sum up to 1, and the ``secondary_sites_vessels`` parameter should be a list of the number of vessels in each secondary grid.
+
+   Thus, both of these parameters should have a length equal to the number of grids - 1. Therefore, we can change the parameters accordingly:
+      
+   .. image:: changing_parameters.png
 
 2. **Start a simulation**: select the ``New Simulation`` option from the main menu.
-3. **Introduce the total time steps for the simulation**: you will be prompted to enter the total number of time steps for the simulation. For this tutorial, we will use 1000 time steps.
+3. **Introduce the total time steps for the simulation**: you will be prompted to enter the total number of time steps that the simulation will last. For this tutorial, we will use 1000 time steps.
 
 .. image:: select_total_timesteps.png
 
 4. **Introduce the interval for saving the results**: you will be prompted to enter the interval for saving the results. This is the number of time steps between each saved result. For this tutorial, we will use an interval of 50 time steps.
 
-.. image:: select_save_interval.png
+.. image:: select_saving_interval.png
 
-This will start the simulation, you will see the progress of the simulation in the terminal. The simulation will run for the specified number of time steps, saving the results at the specified interval:
-.. image:: simulation_progress.png
[…]
