+++ docs/source/usage.rst
@@ -1,6 +1,0 @@
-Usage
-=====
-
-Example of autofunction:
-
-.. autofunction:: metaspread.CancerModel
