### 2025-07-21 edb0ad9
delete cli tutorial from toctree
