### 2025-07-21 c69d5ef
fix table reference in introduction
