### 2025-07-21 eaf14f3
add contents table to tutorial interactive
