+++ docs/source/tutorial-interactive.rst
@@ -1,8 +1,8 @@
 Tutorial: running a simulation in interactive mode
 =====
 
-.. :contents::
-    :depth: 1
+.. contents::
+   :depth: 1
 
 This tutorial will guide you through the basic usage of MetaSpread through the interactive mode, from running a simulation to visualizing the results.
 
