### 2025-07-21 a292236
add quickstart to readme
