### 2025-07-21 a107f4b
update quickstart on readme
