### 2025-07-21 alfredohernandezinostroza/MetaSpread#1 @alfredohernandezinostroza
Hi, thank you for checking on the updated version of the paper!

We have added a paragraph of the section titled “Directions for extension and links with other existing packages.” Please let me know any further adjustments are needed.

Thank you again for your suggestions!
