### 2025-07-21 fb0200d
update quickstart and add gif
