### 2025-07-21 53687b4
add cli tutorial
