### 2025-07-21 7b88947
fix errors in interactive tutorial
