### 2025-07-21 bd1dc95
fix color highlighting for readme
