### 2025-07-21 2837982
update depth contents table to 2 on both tutorials
