### 2025-07-21 097b53f
fix errors in cli tutorial
