### 2025-07-21 f7727e3
fix tutorial interactive
