### 2025-07-21 3e51f2c
update interactive tutorial
