### 2025-07-21 #8092 @alfredohernandezinostroza
@majensen @tushardave26 @chryswoods

Hello! Thank you for your patience. I apologize for the delay in my response.

I have updated the documentation and made the required changes to the README. Regarding the documentation improvements, I have I added two tutorials that will help the users get started with the package.

And concerning the paper, we have added a new paragraph on the section "Directions for extension and links with other existing packages."

Thank you very much for your support and feedback throughout this process. Please let me know if you have any further comments or suggestions!
