### 2025-07-21 e20c497
update cli tutorial
