+++ docs/source/tutorial-cli.rst
@@ -49,10 +49,10 @@ Running a Simulation
 
    .. code-block:: console
 
-      python -m metaspread postprocess all Sim-max_steps-300-collection_period-30-cells-388-grids_number-3 10 1
+      python -m metaspread postprocess all Sim-max_steps-300-collection_period-30-cells-388-grids_number-2 10 1
 
 4. **Visualize the results**
-   Go to the `Videos` inside the `Sim-max_steps-300-collection_period-30-cells-388-grids_number-3` directory and check the results. You should get something like the following:
+   Go to the `Videos` inside the `Sim-max_steps-300-collection_period-30-cells-388-grids_number-2` directory and check the results. You should get something like the following:
 
    .. image:: video_tumor_dynamics_cli.gif
       :align: center
