### 2025-07-21 7cc8de1
update cli tutorial
