+++ docs/source/tutorial-cli.rst
@@ -51,8 +51,7 @@ Running a Simulation
 
       python -m metaspread postprocess all Sim-max_steps-300-collection_period-30-cells-388-grids_number-2 10 1
 
-4. **Visualize the results**
-   Go to the `Videos` inside the `Sim-max_steps-300-collection_period-30-cells-388-grids_number-2` directory and check the results. You should get something like the following:
+4. **Visualize the results**: Enter the `Videos` directory inside the `Sim-max_steps-300-collection_period-30-cells-388-grids_number-2` directory and check the results. You should get something like the following:
 
    .. image:: video_tumor_dynamics_cli.gif
       :align: center
