### 2025-07-31 #8092 @chryswoods
I confirm that I have been able to download and successfully run the new version. The simulation runs as expected, and the tests all appear to pass on MacOS 15.6 using Python 3.12.

I think it is nearly there now. The only comments I have are:

1. You say that Python 3.10 is the only version supported. Python 3.10 is end of life next year (https://devguide.python.org/versions/). I've run using 3.12, which works fine. Could you test this yourself, and maybe say that Python 3.12 or Python 3.13 are supported?

2. I couldn't find any instructions on how to run your unit tests. To run, I had to install pytest and pytest-mock. Could you add a small section in the README telling people how to run the tests so that they can verify everything is working. This is important, I believe, because your quick start example doesn't appear to do much (at least when I ran it). There's little to no growth in the run, and most of the videos are zero lines with just an advancing day. This makes it difficult to judge from the video if the code is "working", or whether this is the expected output for a short simulation?
