### 2025-08-14 1f1fe6c
add testing to README
