### 2025-10-22 #8092 @alfredohernandezinostroza
@majensen @rkurchin
Hello! Done! Just some caveats:
I noticed that in line 78 I also had to change focused for focuses. Furthermore, I improved the wording a little bit:

CancerSim focused on cell heterogeneity and evolution processes with their effect on tumor growth, and MetaSpread focuses on cell population processes
->
CancerSim focuses on cell heterogeneity and evolutionary processes and their effects on tumor growth, while MetaSpread focuses on cell population processes.

Finally, l.121 and l.124 cannot be corrected. I assume that during the compilation process the capitalization of the words is corrected automatically, as I had already written Python and Markov (you can check it on paper.bib). To test this, I changed Markov to MArkov, and it did not change anything in the PDF.

Let me know if there is anything else I should do on my end!
